spider_django/requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
Django==3.1.3
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.2
jmespath==0.10.0
psycopg2==2.8.6
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
s3transfer==0.3.3
screen==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
xlwt==1.3.0
